--- EscapeCOVIDConsole/Window.cpp
@@ -1,7 +1,0 @@
-#include "Window.h"
-
-Window::Window()
-{
-    //RenderWindow window(VideoMode(FRAME_WIDTH, FRAME_LENGTH),
-    //    "EscapeCOVID SFML C++");//build window
-}
