--- EscapeCOVIDConsole/Window.h
@@ -1,9 +1,0 @@
-#ifndef WINDOW_H
-#define WINDOW_H
-#include "rootheader.h"
-class Window
-{
-    Window();
-};
-
-#endif WINDOW_H
