+++ EscapeCOVIDConsole/virusObj.h
